+++ app/build.gradle
@@ -8,8 +8,8 @@ android {
         applicationId "io.github.anotherme17.commonadaptersample"
         minSdkVersion 15
         targetSdkVersion 23
-        versionCode 2
-        versionName "1.2.0"
+        versionCode 4
+        versionName "0.4.1"
     }
     buildTypes {
         release {
