+++ common-rv-adapter/build.gradle
@@ -9,8 +9,8 @@ android {
     defaultConfig {
         minSdkVersion 15
         targetSdkVersion 23
-        versionCode 2
-        versionName "1.2.0"
+        versionCode 4
+        versionName "0.4.1"
     }
     buildTypes {
         release {
