--- common-rv-adapter/src/main/java/io/github/anotherme17/commonrvadapter/adapter/RvHeadAndFootAdapter.java
@@ -1,139 +1,0 @@
-package io.github.anotherme17.commonrvadapter.adapter;
-
-import android.support.v4.util.SparseArrayCompat;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
-import android.view.View;
-import android.view.ViewGroup;
-
-/**
- * Created by user798 on 2016/12/28.
- */
-
-public  class RvHeadAndFootAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
-
-    private static final int BASE_ITEM_TYPE_HEADER = 1024;
-    private static final int BASE_ITEM_TYPE_FOOTER = 2048;
-
-    private SparseArrayCompat<View> mHeaderViews = new SparseArrayCompat<>();
-    private SparseArrayCompat<View> mFootViews = new SparseArrayCompat<>();
-
-    private RecyclerView.Adapter mInnerAdapter;
-
-    public RvHeadAndFootAdapter(RecyclerView.Adapter adapter) {
-        this.mInnerAdapter = adapter;
-    }
-
-    public RecyclerView.Adapter getInnerAdapter() {
-        return mInnerAdapter;
-    }
-
-    @Override
-    public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-
-        if (mHeaderViews.get(viewType) != null) {
-            return new RecyclerView.ViewHolder(mHeaderViews.get(viewType)) {
-            };
-        } else if (mFootViews.get(viewType) != null) {
-            return new RecyclerView.ViewHolder(mFootViews.get(viewType)) {
-            };
-        }
-        return mInnerAdapter.createViewHolder(parent, viewType);
-    }
-
-    @Override
-    public int getItemViewType(int position) {
-        if (isHeaderView(position)) {
-            return mHeaderViews.keyAt(position);
-        } else if (isFooterView(position)) {
-            return mFootViews.keyAt(position - getHeadersCount() - getRealItemCount());
-        }
-        return mInnerAdapter.getItemViewType(position - getHeadersCount());
-    }
-
-    @Override
-    public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
-        if (isHeaderView(position)) {
-            return;
-        }
-        if (isFooterView(position)) {
-            return;
-        }
-        mInnerAdapter.onBindViewHolder(holder, position - getHeadersCount());
-    }
-
-    @Override
-    public int getItemCount() {
-        return getHeadersCount() + getFootersCount() + getRealItemCount();
-    }
-
-    @Override
-    public void onAttachedToRecyclerView(RecyclerView recyclerView) {
-        mInnerAdapter.onAttachedToRecyclerView(recyclerView);
-
-        RecyclerView.LayoutManager layoutManager = recyclerView.getLayoutManager();
-        if (layoutManager instanceof GridLayoutManager) {
-            final GridLayoutManager gridLayoutManager = (GridLayoutManager) layoutManager;
-            final GridLayoutManager.SpanSizeLookup spanSizeLookup = gridLayoutManager.getSpanSizeLookup();
-
-            gridLayoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
-                @Override
-                public int getSpanSize(int position) {
-                    int viewType = getItemViewType(position);
-                    if (mHeaderViews.get(viewType) != null) {
-                        return gridLayoutManager.getSpanCount();
-                    } else if (mFootViews.get(viewType) != null) {
-                        return gridLayoutManager.getSpanCount();
-                    }
-                    if (spanSizeLookup != null) {
-                        return spanSizeLookup.getSpanSize(position - getHeadersCount());
-                    }
-                    return 1;
-                }
-            });
-        }
-    }
-
-    @Override
-    public void onViewAttachedToWindow(RecyclerView.ViewHolder holder) {
-        mInnerAdapter.onViewAttachedToWindow(holder);
-        int position = holder.getLayoutPosition();
-        if (isHeaderView(position) || isFooterView(position)) {
-            ViewGroup.LayoutParams lp = holder.itemView.getLayoutParams();
-            if (lp != null && lp instanceof StaggeredGridLayoutManager.LayoutParams) {
-                StaggeredGridLayoutManager.LayoutParams p = (StaggeredGridLayoutManager.LayoutParams) lp;
-                p.setFullSpan(true);
-            }
-        }
-    }
-
-
-    public void addHeaderView(View view) {
-        mHeaderViews.put(mHeaderViews.size() + BASE_ITEM_TYPE_HEADER, view);
-    }
-
-    public void addFooterView(View view) {
-        mFootViews.put(mFootViews.size() + BASE_ITEM_TYPE_FOOTER, view);
-    }
-
-    private int getRealItemCount() {
-        return mInnerAdapter.getItemCount();
-    }
-
-    public int getHeadersCount() {
-        return mHeaderViews.size();
-    }
-
-    public int getFootersCount() {
-        return mFootViews.size();
-    }
-
-    private boolean isHeaderView(int position) {
-        return position < getHeadersCount();
-    }
-
-    private boolean isFooterView(int position) {
-        return position >= getHeadersCount() + getRealItemCount();
-    }
-}
